Move executor creation to own file
